1/04
Give move now works as intended. After this function was implemented I realised that a Remove move function would also be necessary and was thus implemented, however an error occurs when the user tries to remove a move from a pokemon that does not exist on their team. This will be remedied next update and the api will finally be implemented.