2/04
All functions bar challenge (as it is essentially an entire other project) have been complete and now be complete and I began implementing the api. It is now used to check if a pokemon actually exists when a user tries to add it to their party in the store pokemon function and in the give move function to determine if the entered move is real and compatible with the pokemon. Next time I will integrate the api into the search function